Configure Gruntfile to run specs correctly
The line: 'grunt.registerTask('test', ['testServer'])' was causing the followng error:

  Loading "Gruntfile.js" tasks...ERROR
  >> TypeError: undefined is not a function

So it has been temporarily removed

Likewise, the task involved in activating the 'watch' sub-task was also causing an error:

  Verifying property watch.client.files exists in config...ERROR
  >> Unable to process task.
  Warning: Required config property "watch.client.files" missing.

I have also replaced the package 'grunt-mocha' with 'grunt-mocha-test' because the
former is a testing framework that is meant for frontend and the latter is meant
for backend. Since we are primarily concerned with testing services that
are being used on the backend I felt using 'grunt-mocha-test' was more
fitting package.

I have removed Karma because Karma is a test runner that is meant for
testing frontend code in psuedo-browsers, and we are focusing primarily
on testing backend services. Karma was used in the Testify sprint
for the purpose of testing client side API calls; I do not feel using
Karma in this project is justified.

I believe the best testing stack for our backend services would be
mocha, chai, and sinon. This commit contains the configurations for
integrating that testing stack with Grunt.